requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
billiard==4.2.1
build==1.2.2.post1
cachetools==5.5.1
celery==5.3.6
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
fastapi==0.115.8
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
idna==3.10
jiter==0.8.2
kombu==5.4.2
mysql-connector-python==9.2.0
numpy==2.2.3
openai==1.61.1
packaging==24.2
parsimonious==0.10.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.2.1
PyPDF2==3.0.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
regex==2024.11.6
requests==2.32.3
rsa==4.9
setuptools==75.8.0
singlestoredb==1.11.0
six==1.17.0
sniffio==1.3.1
sqlparams==6.2.0
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wheel==0.45.1
